requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy
pandas
opencv-python-headless
albumentations==1.3.1
streamlit
pydicom
scikit-image
pytest
flake8
python-dotenv

kaggle
